requirements.txt
Django==4.2.16
django-widget-tweaks==1.5.0
pillow==11.3.0
reportlab==4.4.3
whitenoise==6.6.0
gunicorn==21.2.0
google-generativeai==0.3.2
requests==2.31.0
